split batchmessage.go
